调整测试设计页面高度
个人概览改为dashboard；隐藏需求中心入口；删除系统公告

测试设计默认展开根节点

注册产品版本产品栏去掉自动完成

批量注册虚拟机修改

feat: workspace后端适配

feat: 补充case查询接口workspace适配

feat: 为非法workspace访问新增4021类报错

fix:convert str error

删除无用代码

顶部菜单修改

顶部菜单新增图标

workspace相关修改

workspace页面修改

fix: 修改workspace适配bug&前端部分样式调整

fix:messenger token authenticate

workspace修改

to: workspaces页面添加团队&用户数统计

fix:job adding fields creator_id

fix:time module error

fix:lack of pytz module

fix:解决图片模糊；图片随机背景色；页脚修改；修复变量拼写错误；

org_id批量替换；用例管理group接口新增参数

修复路由跳转caseId参数错误

fix: 接口常规装饰器遗漏补充&机器组证书主动失效接口安全加固

fix:time error

fix:vmachine des error

fix:manual job change fileds executor_id to creator_id

fix:template return data change

修复个人积分/团队积分分页BUG

Sockets等上限由2改为4

修改执行表单传参

manual-job接口修改

修复创建模板无法选中用例的bug

fix:remove login cookie info to data

删除cookie

修复虚拟机批量创建bug

修复bug

删除refresh_token

fix:add messenger serializer

fix:logger info messenger token

fix:messenger token secret remove

fix:issue state log

fix:remove issue state log

修复测试进展测试进度条显示BUG

质量checklist放入表格操作栏

优化产品版本编辑操作

修复创建模板BUG

修复测试执行创建模板BUG

报错弹框限时

右上角报错最多显示3个

修复用例管理菜单树右键无效错误

fix:executor_id to creator_id

fix: 修复版本基线删除BUG&删除CaseNode表冗余关联关系

feat: 新增为用例集绑定游离测试套的入口

feat: 新增LTP用例解析适配&&GitRepo表新增branch字段

fix: remove permission type limitation while adding new GitRepo data

to: 修复仓库解析shlex导入错误&&仓库解析加锁&&仓库解析新增手动触发入口

fix: 修复scriptLoader键值存储错误&优化手动同步前端提示

fix: 代码仓手动同步异步任务补充userid记录&&优化异常处理

fix: 优化gitClone间隔

to: 去除仓库同步循环gitClone逻辑&&补全日志打印

fix: build-in问题修复

fix: git命令错误

fix: 仓库同步最终debug

fix: 修复测试套数量统计因缺少筛选导致的计数错误

fix: 用例管理团队数据范围修正&&GitRepo代码仓地址字段改为非唯一

to: 拆分framework与gitrepo接口文件

to: 代码仓解析入参减少&&定时同步逻辑修改

feat: 批量新增节点采用celery异步拆分子任务提高接口性能

fix: try代码块包裹范围优化

to: casenode创建异步任务编写优化

fix:register url bug

fix: 代码仓同步逻辑判断错误修正

fix: 删除executor适配器中冗余定义的接口

feat: adding the adaptors of repo resolvers

fix: 补充缺少的add update

to: 取消定时仓库同步以屏蔽脏数据问题

fix: 修复递归获取目录下用例文件逻辑分支异常

+++ radiaTest-server/server/apps/testcase/handler.py
@@ -261,8 +261,8 @@ class CaseNodeHandler:
 
     @staticmethod
     @collect_sql_error
-    def get_roots(query):
-        filter_params = GetAllByPermission(CaseNode).get_filter()
+    def get_roots(query, workspace=None):
+        filter_params = GetAllByPermission(CaseNode, workspace).get_filter()
         filter_params.append(CaseNode.is_root.is_(True))
         for key, value in query.dict().items():
             if not value:
@@ -371,12 +371,15 @@ class CaseNodeHandler:
 
         if current_org_id != case_node.org_id:
             return jsonify(error_code=RET.VERIFY_ERR, error_msg="No right to delete")
-        if case_node.type == "baseline":
-            baseline = Baseline.query.filter_by(id=case_node.baseline_id).first()
-            db.session.delete(baseline)
-        else:
-            db.session.delete(case_node)
 
+        # 解决级联删除未生效问题
+        _baseline_id = case_node.baseline_id
+        db.session.delete(case_node)
+        
+        if case_node.type == "baseline" and _baseline_id:
+            baseline = Baseline.query.filter_by(id=_baseline_id).first()
+            db.session.delete(baseline)
+            
         try:
             db.session.commit()
         except (IntegrityError, SQLAlchemyError) as e:
@@ -1304,6 +1307,8 @@ class ResourceItemHandler:
 
     def get_case(self):
         cases_filter = self._get_table_filter(Case)
+        cases_filter.append(CaseNode.type == 'case')
+
         if self.case_node_type == 'baseline':
             cases_filter.append(CaseNode.baseline_id == self.baseline_id)
         else:
@@ -1337,6 +1342,8 @@ class ResourceItemHandler:
 
     def get_suite(self):
         suites_filter = self._get_table_filter(Suite)
+        suites_filter.append(CaseNode.type == 'suite')
+
         if self.case_node_type == 'baseline':
             suites_filter.append(CaseNode.baseline_id == self.baseline_id)
         else:
@@ -1558,3 +1565,55 @@ class SuiteDocumentHandler:
         _id = Insert(SuiteDocument, _body).insert_id(SuiteDocument, "/suite_document")
         return jsonify(error_code=RET.OK, error_msg="OK", data={"id": _id})
 
+
+class OrphanSuitesHandler:
+    def __init__(self, query) -> None:
+        self.current_org_id = redis_client.hget(
+            RedisKey.user(g.gitee_id),
+            "current_org_id",
+        )
+
+        self.filter = [
+            Suite.org_id == int(self.current_org_id),
+            Suite.permission_type == "org",
+            CaseNode.id == sqlalchemy.null(),
+        ]
+
+        self.query = query
+        self._add_params()
+
+    def _add_params(self):
+        _filter_params = []
+        if self.query.name:
+            _filter_params.append(Suite.name.like(f"%{self.query.name}%"))
+        if self.query.owner:
+            _filter_params.append(Suite.owner == self.query.owner)
+        if self.query.git_repo_url:
+            _filter_params.append(Suite.git_repo.git_url == self.query.git_repo_url)
+        if self.query.framework_name:
+            _filter_params.append(Suite.framework.name == self.query.framework_name)
+        self.filter += _filter_params
+
+    def add_filters(self, query_filters: list):
+        self.filter += query_filters
+
+    @collect_sql_error
+    def get_all(self):
+        query_filter = Suite.query.outerjoin(CaseNode).filter(*self.filter).order_by(
+            Suite.name, 
+            Suite.create_time
+        )
+
+        def page_func(item):
+            suite_dict = item.to_json()
+            return suite_dict
+        
+        page_dict, e = PageUtil.get_page_dict(
+            query_filter, 
+            self.query.page_num, 
+            self.query.page_size, 
+            func=page_func
+        )
+        if e:
+            return jsonify(error_code=RET.SERVER_ERR, error_msg=f'get orphan suites page error {e}')
+        return jsonify(error_code=RET.OK, error_msg="OK", data=page_dict)
